requirements.txt
telegram
kinopoiskpy
python-telegram-bot